+++ src/gui/activitywidget.h
@@ -22,6 +22,7 @@
 #include "progressdispatcher.h"
 #include "owncloudgui.h"
 #include "account.h"
+#include "activitydata.h"
 
 #include "ui_activitywidget.h"
 
@@ -35,119 +36,12 @@ class AccountStatusPtr;
 class ProtocolWidget;
 class JsonApiJob;
 class NotificationWidget;
+class ActivityListModel;
 
 namespace Ui {
   class ActivityWidget;
 }
 class Application;
-
-/**
- * @brief The ActivityLink class describes actions of an activity
- *
- * These are part of notifications which are mapped into activities.
- */
-
-class ActivityLink
-{
-public:
-    QHash <QString, QVariant> toVariantHash();
-
-    QString _label;
-    QString _link;
-    QString _verb;
-    bool _isPrimary;
-};
-
-/**
- * @brief Activity Structure
- * @ingroup gui
- *
- * contains all the information describing a single activity.
- */
-
-class Activity
-{
-public:
-    enum Type {
-        ActivityType,
-        NotificationType
-    };
-    Type      _type;
-    qlonglong _id;
-    QString   _subject;
-    QString   _message;
-    QString   _file;
-    QUrl      _link;
-    QDateTime _dateTime;
-    QString   _accName;
-
-    QVector <ActivityLink> _links;
-    /**
-     * @brief Sort operator to sort the list youngest first.
-     * @param val
-     * @return
-     */
-    bool operator<( const Activity& val ) const {
-        return _dateTime.toMSecsSinceEpoch() > val._dateTime.toMSecsSinceEpoch();
-    }
-
-};
-
-/**
- * @brief The ActivityList
- * @ingroup gui
- *
- * A QList based list of Activities
- */
-class ActivityList:public QList<Activity>
-{
-public:
-    void setAccountName( const QString& name );
-    QString accountName() const;
-
-private:
-    QString _accountName;
-};
-
-
-/**
- * @brief The ActivityListModel
- * @ingroup gui
- *
- * Simple list model to provide the list view with data.
- */
-class ActivityListModel : public QAbstractListModel
-{
-    Q_OBJECT
-public:
-    explicit ActivityListModel(QWidget *parent=0);
-
-    QVariant data(const QModelIndex &index, int role) const Q_DECL_OVERRIDE;
-    int rowCount(const QModelIndex& parent = QModelIndex()) const Q_DECL_OVERRIDE;
-
-    bool canFetchMore(const QModelIndex& ) const Q_DECL_OVERRIDE;
-    void fetchMore(const QModelIndex&) Q_DECL_OVERRIDE;
-
-    ActivityList activityList() { return _finalList; }
-
-public slots:
-    void slotRefreshActivity(AccountState* ast);
-    void slotRemoveAccount( AccountState *ast );
-
-private slots:
-    void slotActivitiesReceived(const QVariantMap& json, int statusCode);
-
-signals:
-    void activityJobStatusCode(AccountState* ast, int statusCode);
-
-private:
-    void startFetchJob(AccountState* s);
-    void combineActivityLists();
-
-    QMap<AccountState*, ActivityList> _activityLists;
-    ActivityList _finalList;
-    QSet<AccountState*> _currentlyFetching;
-};
 
 /**
  * @brief The ActivityWidget class
